Big cleanup that moves functions where they belong and kill some utility funvtions that were duplicate code
